Move ConvertToString FormatHelper
