+++ src/NLog/NestedDiagnosticsContext.cs
@@ -56,7 +56,7 @@ namespace NLog
         {
             get
             {
-                return GlobalDiagnosticsContext.ConvertToString(TopObject, null);
+                return FormatHelper.ConvertToString(TopObject, null);
             }
         }
 
@@ -117,7 +117,7 @@ namespace NLog
         /// <returns>The top message, which is removed from the stack, as a string value.</returns>
         public static string Pop(IFormatProvider formatProvider)
         {
-            return GlobalDiagnosticsContext.ConvertToString(PopObject(), formatProvider);
+            return FormatHelper.ConvertToString(PopObject(), formatProvider);
         }
 
         /// <summary>
@@ -154,7 +154,7 @@ namespace NLog
         /// <returns>Array of strings.</returns>
         public static string[] GetAllMessages(IFormatProvider formatProvider) 
         {
-            return ThreadStack.Select((o) => GlobalDiagnosticsContext.ConvertToString(o, formatProvider)).ToArray();
+            return ThreadStack.Select((o) => FormatHelper.ConvertToString(o, formatProvider)).ToArray();
         }
 
         /// <summary>
